(C) and move files around
