+++ src/DI/ServiceLookup/IServiceProviderEngine.cs
@@ -1,7 +1,9 @@
-using System;
-using Microsoft.Extensions.DependencyInjection.ServiceLookup;
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-namespace Microsoft.Extensions.DependencyInjection
+using System;
+
+namespace Microsoft.Extensions.DependencyInjection.ServiceLookup
 {
     internal interface IServiceProviderEngine : IDisposable
     {
