+++ src/DI/ServiceLookup/ServiceProviderEngine.cs
@@ -1,11 +1,13 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
 using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
-using Microsoft.Extensions.DependencyInjection.ServiceLookup;
 
-namespace Microsoft.Extensions.DependencyInjection
+namespace Microsoft.Extensions.DependencyInjection.ServiceLookup
 {
     internal class ServiceProviderEngine : IServiceProviderEngine, IServiceScopeFactory
     {
