+++ src/DI/ServiceLookup/ServiceProviderEngineScope.cs
@@ -1,7 +1,10 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
 using System;
 using System.Collections.Generic;
 
-namespace Microsoft.Extensions.DependencyInjection
+namespace Microsoft.Extensions.DependencyInjection.ServiceLookup
 {
     internal class ServiceProviderEngineScope : IServiceScope, IServiceProvider
     {
